updated files

+++ src/app/employee-list/employee-list.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { ConfirmationDialogueComponent } from '../confirmation-dialogue/confirmation-dialogue.component';
 import { Employee} from '../employee'
 import { EmployeeserviceService } from '../employeeservice.service';
 @Component({
@@ -9,7 +11,7 @@ import { EmployeeserviceService } from '../employeeservice.service';
 })
 export class EmployeeListComponent implements OnInit {
 employees!: Employee[];
-  constructor(private employeeser:EmployeeserviceService,private router: Router) { }
+  constructor(private employeeser:EmployeeserviceService,private router: Router,public dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.getEmployees();
@@ -26,13 +28,43 @@ employees!: Employee[];
     this.router.navigate(['employee-details', id]);
   }
   updateEmployee(id: number){
-    this.router.navigate(['update-employee',id]);
-    console.log(id);
-  }
-  deleteEmployee(id: number){
-    this.employeeser.deleteEmployee(id).subscribe( data => {
-      console.log(data);
-      this.getEmployees();
-    })
+    const confirmDialog = this.dialog.open(ConfirmationDialogueComponent, {
+      data: {
+        title: 'Confirm Remove Employee',
+        message: 'Are you sure, you want to Update an employee: '
+      }
+    });
+    confirmDialog.afterClosed().subscribe(result => {
+      if (result === true) {
+        this.router.navigate(['update-employee',id]);
+        console.log(id);
+      }
+        });
+      }
+   
+  // deleteEmployee(id: number){
+  //   alert("Are you sure You want to delete?");
+  //   this.employeeser.deleteEmployee(id).subscribe( data => {
+  //     console.log(data);
+  //     this.getEmployees();
+  //   })
+  // }
+ 
+  deleteEmployee(id: number) {
+    const confirmDialog = this.dialog.open(ConfirmationDialogueComponent, {
+      data: {
+        title: 'Confirm Remove Employee',
+        message: 'Are you sure, you want to remove an employee: '
+      }
+    });
+    confirmDialog.afterClosed().subscribe(result => {
+      if (result === true) {
+        this.employeeser.deleteEmployee(id).subscribe( data => {
+          console.log(data);
+          this.getEmployees();
+        })
+      }
+    });
   }
 }
+
